Added markdown support and sanitized it.
