+++ package.json
@@ -15,6 +15,7 @@
 		"@melt-ui/pp": "^0.0.7",
 		"@sveltejs/adapter-auto": "^2.0.0",
 		"@sveltejs/kit": "^1.20.4",
+		"@tailwindcss/typography": "^0.5.9",
 		"@typescript-eslint/eslint-plugin": "^5.45.0",
 		"@typescript-eslint/parser": "^5.45.0",
 		"autoprefixer": "^10.4.14",
@@ -41,7 +42,9 @@
 		"@fontsource-variable/inter": "^5.0.5",
 		"@melt-ui/svelte": "^0.27.0",
 		"@types/node": "^20.4.5",
+		"isomorphic-dompurify": "^1.8.0",
 		"nanoid": "^4.0.2",
+		"snarkdown": "^2.0.0",
 		"socket.io-client": "^4.7.1",
 		"unique-names-generator": "^4.7.1",
 		"valibot": "^0.3.0"
