+++ src/lib/components/messages/CreateMessageForm.svelte
@@ -1,42 +1,9 @@
 <script lang="ts">
-	import { DiceRoll } from '@dice-roller/rpg-dice-roller';
-	import { user } from '$lib/stores/user-store';
-	import { onMount } from 'svelte';
-	import type {
-		CreateMessageEvent,
-		InvalidDataEvent,
-		MessageCreatedEvent
-	} from '$lib/types/socket-types';
-	import { validCreateMessage, validCreateRoom } from '$lib/validation/processors';
 	import { socket } from '$lib/sockets/client';
-	import { messages } from '$lib/stores/messages-store';
-
-	function handleErrorMessage(message: string) {
-		console.error(message);
-	}
-
-	function onMessageCreated(event: MessageCreatedEvent) {
-		console.debug(event);
-		event.message.createdAt = new Date(event.message.createdAt);
-		$messages = $messages ? [...$messages, event.message] : [event.message];
-	}
-
-	function onInvalidData(event: InvalidDataEvent) {
-		console.debug(event);
-		handleErrorMessage(event?.message);
-	}
-
-	onMount(() => {
-		socket.on('message-created', onMessageCreated);
-		socket.on('invalid-data', onInvalidData);
-
-		return () => {
-			console.debug(`socket-off:message-created:${socket.id}`);
-			console.debug(`socket-off:invalid-data:${socket.id}`);
-			socket.off('message-created', onMessageCreated);
-			socket.off('invalid-data', handleErrorMessage);
-		};
-	});
+	import { user } from '$lib/stores/user-store';
+	import type { CreateMessageEvent } from '$lib/types/socket-types';
+	import { validCreateMessage } from '$lib/validation/processors';
+	import { DiceRoll } from '@dice-roller/rpg-dice-roller';
 
 	function submitOnEnter(e: KeyboardEvent) {
 		if (e.key === 'Enter' && !e.shiftKey) {
